pyproject.toml
[tool.poetry]
name = "homework_31"
version = "0.1.0"
description = ""
authors = ["Кристина Ратник <kristi.2001@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1"
psycopg2 = "^2.9.3"
Pillow = "^9.2.0"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
